goals are protected with user JWT

+++ backend/controllers/goalControl.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel'); 
 
 // @desc GET goals
 // @route /api/goals/
@@ -35,6 +36,18 @@ const updateGoal = asyncHandler (async (req, res) =>{
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+    //Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure that the logged in user matches the goal created user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -50,6 +63,20 @@ const deleteGoal = asyncHandler (async (req, res) =>{
         res.status(404)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure that the logged in user matches the goal created user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await goal.deleteOne()
     res.status(200).json({id: req.params.id});
 })
